Made the project buildable.

+++ arrus/core/devices/us4r/external/ius4oem/DTGCAttenuationValueMap.h
@@ -12,7 +12,7 @@ namespace arrus::devices {
 class DTGCAttenuationValueMap {
 
 public:
-    using ValueType = uint8;
+    using ValueType = uint16;
 
     static DTGCAttenuationValueMap &getInstance() {
         static DTGCAttenuationValueMap instance;
